+++ Livrable1.logger/bin/Debug/net8.0/Livrable1.logger.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v8.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v8.0": {
+      "Livrable1.logger/1.0.0": {
+        "runtime": {
+          "Livrable1.logger.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "Livrable1.logger/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
