+++ Livrable1/Logs/backup_state.json
@@ -1,0 +1,22 @@
+{
+  "Backup Operation": {
+    "BackupName": "Backup Operation",
+    "LastActionTimestamp": "2025-02-11T16:13:10.1169007+01:00",
+    "CurrentAction": "All backups completed"
+  },
+  "ACIDE": {
+    "BackupName": "ACIDE",
+    "LastActionTimestamp": "2025-02-11T16:08:07.1996505+01:00",
+    "CurrentAction": "Backup completed successfully"
+  },
+  "Ttetst": {
+    "BackupName": "Ttetst",
+    "LastActionTimestamp": "2025-02-11T16:11:10.1888408+01:00",
+    "CurrentAction": "Backup completed successfully"
+  },
+  "LOGS": {
+    "BackupName": "LOGS",
+    "LastActionTimestamp": "2025-02-11T16:13:10.1159657+01:00",
+    "CurrentAction": "Backup completed successfully"
+  }
+}
