+++ Livrable1/Logs/state.json
@@ -1,12 +1,34 @@
 [
   {
-    "name": "test",
-    "realDirectoryPath": "C:\\ProjetEasySave\\path\\source",
-    "copyDirectoryPath": "C:\\ProjetEasySave\\path\\destination",
+    "name": "ACIDE",
+    "realDirectoryPath": "C:\\Users\\picou\\Documents\\FISA-INFO-A3\\GENIE_LOGICIEL\\ProjetV2\\G2_EasySave\\clonage\\Source",
+    "copyDirectoryPath": "C:\\Users\\picou\\Documents\\FISA-INFO-A3\\GENIE_LOGICIEL\\ProjetV2\\G2_EasySave\\clonage\\Destination",
+    "fileNumber": 2,
+    "dernierHorodatage": "2025-02-11T16:08:07.1982716+01:00",
+    "actif": true,
+    "tailleTotale": 0,
+    "fichiersRestants": 0,
+    "tailleRestante": 0
+  },
+  {
+    "name": "Ttetst",
+    "realDirectoryPath": "C:\\Users\\picou\\Documents\\FISA-INFO-A3\\GENIE_LOGICIEL\\ProjetV2\\G2_EasySave\\clonage\\Source",
+    "copyDirectoryPath": "C:\\Users\\picou\\Documents\\FISA-INFO-A3\\GENIE_LOGICIEL\\ProjetV2\\G2_EasySave\\clonage\\Destination",
     "fileNumber": 1,
-    "dernierHorodatage": "2025-02-10T09:09:10.1753715+01:00",
-    "actif": false,
-    "tailleTotale": 967,
+    "dernierHorodatage": "2025-02-11T16:11:10.1873333+01:00",
+    "actif": true,
+    "tailleTotale": 0,
+    "fichiersRestants": 0,
+    "tailleRestante": 0
+  },
+  {
+    "name": "LOGS",
+    "realDirectoryPath": "C:\\Users\\picou\\Documents\\FISA-INFO-A3\\GENIE_LOGICIEL\\ProjetV2\\G2_EasySave\\clonage\\Source",
+    "copyDirectoryPath": "C:\\Users\\picou\\Documents\\FISA-INFO-A3\\GENIE_LOGICIEL\\ProjetV2\\G2_EasySave\\clonage\\Destination",
+    "fileNumber": 1,
+    "dernierHorodatage": "2025-02-11T16:13:10.1144699+01:00",
+    "actif": true,
+    "tailleTotale": 0,
     "fichiersRestants": 0,
     "tailleRestante": 0
   }
