--- Livrable1/bin/Debug/net8.0/Logs/backup_state.json
@@ -1,17 +1,0 @@
-{
-  "Backup Operation": {
-    "BackupName": "Backup Operation",
-    "LastActionTimestamp": "2025-02-09T18:12:51.9800498+01:00",
-    "CurrentAction": "All backups completed"
-  },
-  "test": {
-    "BackupName": "test",
-    "LastActionTimestamp": "2025-02-09T18:11:55.6995135+01:00",
-    "CurrentAction": "Sauvegarde termin\u00E9e avec succ\u00E8s"
-  },
-  "logstest": {
-    "BackupName": "logstest",
-    "LastActionTimestamp": "2025-02-09T18:12:51.9792335+01:00",
-    "CurrentAction": "Sauvegarde termin\u00E9e avec succ\u00E8s"
-  }
-}
