--- Livrable1/bin/Debug/net8.0/Logs/log_2025-02-07.json
@@ -1,10 +1,0 @@
-[
-  {
-    "Name": "test",
-    "FileSource": "D:\\clonage\\source\\test.txt",
-    "FileTarget": "D:\\clonage\\destination\\test\\test.txt",
-    "FileSize": 4,
-    "FileTransferTime": 0.001,
-    "time": "07/02/2025 19:06:22"
-  }
-]
