--- Livrable1/bin/Debug/net8.0/Logs/log_2025-02-09.json
@@ -1,10 +1,0 @@
-[
-  {
-    "Name": "logstest",
-    "FileSource": "D:\\clonage\\source\\test.txt",
-    "FileTarget": "D:\\clonage\\destination\\logstest\\test.txt",
-    "FileSize": 4,
-    "FileTransferTime": 0,
-    "time": "09/02/2025 18:12:51"
-  }
-]
